# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/public/social-media/{subdomain}": {
            "get": {
                "description": "show list of Public Social Media",
                "tags": [
                    "Public"
                ],
                "summary": "Get List Public Social Media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "subdomain",
                        "name": "subdomain",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "referral",
                        "name": "referral",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/sales/social-media": {
            "get": {
                "description": "show list of Social Media",
                "tags": [
                    "Data Social Media"
                ],
                "summary": "Get List Social Media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "add data of Social Media",
                "tags": [
                    "Data Social Media"
                ],
                "summary": "Add Data Social Media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AddSocialMediaReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/sales/social-media/{id}": {
            "get": {
                "description": "show detail of Public Social Media",
                "tags": [
                    "Data Social Media"
                ],
                "summary": "Get Detail Social Media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Update Social Media",
                "tags": [
                    "Data Social Media"
                ],
                "summary": "Update Social Media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "description": "body payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateSocialMediaSales"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/testimony": {
            "get": {
                "description": "show detail of Customer",
                "tags": [
                    "Data Testimoni"
                ],
                "summary": "Get Detail Customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Update Data of Testimoni",
                "tags": [
                    "Data Testimoni"
                ],
                "summary": "Update Data Testimoni",
                "parameters": [
                    {
                        "description": "body payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Testimoni"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "create data of Testimoni",
                "tags": [
                    "Data Testimoni"
                ],
                "summary": "Create Data Testimoni",
                "parameters": [
                    {
                        "description": "body payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Testimoni"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/testimony/list": {
            "get": {
                "description": "show list of Customer",
                "tags": [
                    "Data Testimoni"
                ],
                "summary": "Get List Customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/testimony/list/{subdomain}/{referral}": {
            "get": {
                "description": "show list of Customer",
                "tags": [
                    "Data Testimoni"
                ],
                "summary": "Get Public public List Customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sub domain",
                        "name": "subdomain",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.AddSocialMediaReq": {
            "type": "object",
            "properties": {
                "data_social_media": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.SocialMediaListReq"
                    }
                }
            }
        },
...
        "request.SocialMediaListReq": {
            "type": "object",
            "properties": {
                "link_embed": {
                    "type": "string"
                },
                "social_media_name": {
                    "type": "string"
                },
                "user_account": {
                    "type": "string"
                }
            }
        },
        "request.Testimoni": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "integer"
                },
                "modified_at": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                },
                "public_access": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdateSocialMediaSales": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "link_embed": {
                    "type": "string"
                },
                "social_media_name": {
                    "type": "string"
                },
                "user_account": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}